third_party/WebKit/LayoutTests/fast/dom/Window/Location/window-shadow-window-using-js-object-with-location-field.html

<html>
<head>
    <script src="../../../../resources/js-test.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    var window = { location: "haxored" };

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
</body>
</html>
